integrate OSM Repository

+++ data/src/main/java/com/example/network/interfaces/OsmService.kt
@@ -10,7 +10,7 @@ interface OsmService {
     @GET("reverse")
     fun reverseGeocode(
         @Query("format") format: String = "json",
-        @Query("lat") latitude: Long,
-        @Query("lon") longitude: Long
+        @Query("lat") latitude: Double,
+        @Query("lon") longitude: Double
     ): Single<OsmResponseDTO>
 }
